actualizando los id de las tablas para usar long en vez de int

+++ Expensive.Domain/Entities/Categories.cs
@@ -2,7 +2,7 @@
 
 public class Categories : BaseEntity
 {
-    public int CategorieId { get; set; }
+    public long CategorieId { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
     public bool IsDefault { get; set; }
